DEV9: Merge _DEV9open into InitNet, and check ethEnable before calling
